+++ spodproc/src/main/Helper.java
@@ -1,0 +1,67 @@
+package main;
+
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.util.ArrayList;
+import java.util.Scanner;
+
+import com.google.gson.Gson;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+
+public class Helper {
+
+    //From https://www.geeksforgeeks.org/how-to-remove-duplicates-from-arraylist-in-java/
+    public static <T> ArrayList<T> removeDuplicates(ArrayList<T> list)
+    {
+  
+        // Create a new ArrayList
+        ArrayList<T> newList = new ArrayList<T>();
+  
+        // Traverse through the first list
+        for (T element : list) {
+  
+            // If this element is not present in newList
+            // then add it
+            if (!newList.contains(element)) {
+  
+                newList.add(element);
+            }
+        }
+  
+        // return the new list
+        return newList;
+    }
+
+    /**
+     * @param fileName
+     * @return a JsonArray of the file
+     **/
+    public static JsonArray readJsonFile(String fileName) {
+        JsonArray jsonArray = new JsonArray();
+        try {
+            File myObj = new File(fileName);
+            Scanner myReader = new Scanner(myObj);
+            StringBuilder sb = new StringBuilder();
+
+            while (myReader.hasNextLine()) {
+                sb.append(myReader.nextLine());
+            }
+
+            Gson gson = new Gson();
+            JsonArray outerArray = gson.fromJson(sb.toString(), JsonArray.class);
+
+            for (JsonElement jsonElement : outerArray) {
+                jsonArray.add(jsonElement.getAsJsonObject());
+            }
+
+            myReader.close();
+            return jsonArray;
+
+        } catch (FileNotFoundException e) {
+            System.out.println("An error occurred.");
+            e.printStackTrace();
+            return null;
+        }
+    }
+}
